refactor(habitat): lift baseline model domain
