+++ tools/habitat-harness/src/service/model/check/policy/structural/execution.policy.ts
@@ -10,6 +10,7 @@ import {
   type HabitatError,
   renderHabitatError,
 } from "@internal/habitat-harness/resources/errors/index";
+import type { BaselineFileSystemPort } from "@internal/habitat-harness/service/model/baseline/index";
 import {
   approvedSourceScanRootsForRules,
   type CheckOptions,
@@ -47,7 +48,6 @@ import {
 } from "@internal/habitat-harness/service/model/rules/policy/catalog.policy";
 import type { RuleSelection } from "@internal/habitat-harness/service/model/rules/policy/selection.policy";
 import { Clock, Effect } from "effect";
-import type { BaselineFileSystemPort } from "../baseline/index.js";
 
 export interface RuleExecutionRecord {
   result: RuleRunResult;
